requirements.txt
flask==2.3.3
werkzeug==2.3.7
langchain==0.0.267
openai==0.28.0
python-dotenv==0.19.0
boto3==1.28.0
chromadb==0.4.0
pypdf==3.9.0
tiktoken==0.5.1
unstructured==0.10.30
pydantic==1.10.16
networkx==3.2.1
sentence-transformers==2.2.2
python-docx==0.8.11
atlassian-python-api==3.41.0
sqlalchemy==2.0.20
prometheus-client==0.17.1
numpy==1.24.3
huggingface-hub==0.14.1